requirements.txt
requests==2.31.0
mysql-connector-python
BeautifulSoup4
tzdata==2023.3
Pillow==10.0.0
pymysql
redis